@@ -1,46 +1,25 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <PackageLicenseExpression>EUPL-1.2</PackageLicenseExpression>
-    <PackageTags>openmod openmod-plugin universal</PackageTags>
-    <RootNamespace>SalaryPlugin</RootNamespace>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <GenerateNugetPackage>true</GenerateNugetPackage>
-    <Product>SalaryPlugin</Product>
-    <Nullable>enable</Nullable>
-    <WarningsAsErrors>nullable</WarningsAsErrors>
-    <NoWarn>$(NoWarn);NU1701;NU1702;CS0436</NoWarn>
-    <LangVersion>9.0</LangVersion>
-    <Version>1.0.5</Version>
-    <InformationalVersion>1.0.5</InformationalVersion>
-    <PackageVersion>1.0.5</PackageVersion>
-    <AssemblyName>SalaryPlugin</AssemblyName>
-    <Authors>Feli</Authors>
-    <Company>FPlugins</Company>
-    <Description>Pay your players for their worktime</Description>
-    <RepositoryUrl>https://github.com/F-Plugins/SalaryPlugin</RepositoryUrl>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFrameworks>netstandard2.0;net461</TargetFrameworks>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<LangVersion>10</LangVersion>
+		<Version>0.0.0</Version>
+		<PackageVersion>0.0.0</PackageVersion>
+		<InformationalVersion>0.0.0</InformationalVersion>
+		<PackageId>Feli.SalaryPlugin</PackageId>
+		<PackageDescription>Simple and working Salary Plugin for openmod</PackageDescription>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<PackageAuthor>Feli</PackageAuthor>
+		<PackageOwners>Feli</PackageOwners>
+		<PackageTags>openmod openmod-plugin universal</PackageTags>
+		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+		<GenerateNugetPackage>true</GenerateNugetPackage>
+	</PropertyGroup>
 
-  <PropertyGroup>
-    <PackageId>Feli.SalaryPlugin</PackageId> <!-- must be unique, should be same as your plugin ID -->
-    <PackageDescription>An openmod plugin that will allow you to Pay your players for their worktime </PackageDescription>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression> <!-- see https://spdx.org/licenses/ -->
-    <PackageAuthor>Feli</PackageAuthor>
-    <PackageTags>openmod openmod-plugin universal</PackageTags> <!-- XXX can be unturned, unityengine or universal depending on your plugin -->
-    <Version>1.0.4</Version> <!-- Your plugins version. Must be semversion, see https://semver.org/ -->
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <GenerateNugetPackage>true</GenerateNugetPackage>  
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="OpenMod.Core" Version="3.0.27+98c507f" />
-    <PackageReference Include="OpenMod.Extensions.Economy.Abstractions" Version="3.0.28" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="config.yaml" />
-    <EmbeddedResource Include="translations.yaml" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="OpenMod.Extensions.Economy.Abstractions" Version="3.5.2" />
+	</ItemGroup>
 
 </Project>
